+++ src/components/Main/Metrics/MetricsContent/MetricsContent.styles.jsx
@@ -1,4 +1,5 @@
 import { styled } from "styled-components";
+import { appred } from "../../../../utils/colors.styles.jsx";
 
 export const $MetricsContent = styled.div`
   display: flex;
@@ -9,5 +10,12 @@ export const $MetricsContent = styled.div`
   row-gap: 20px;
   column-gap: 20px;
 `;
+
+export const $ErrorLabel = styled.div`
+  color: ${appred};
+  text-align: center;
+  font-size: 14px;
+  margin: 35px;
+`
 
 export default $MetricsContent;